+++ ethercrab-wire-derive/src/lib.rs
@@ -25,6 +25,10 @@
 //!   How many bytes this field consumes on the wire. These attributes may not be present at the
 //!   same time.
 //!
+//!   All builtin `u*` and `i*` types do not need this attribute by default; their size is computed
+//!   automatically if it's omitted. If a different bit width is desired, the `#[wire]` attribute is
+//!   required.
+//!
 //! - `#[wire(pre_skip = N)]` OR `#[wire(pre_skip_bytes = N)]`
 //!
 //!   Skip one or more whole bytes before or after this field in the packed representation.
@@ -74,11 +78,11 @@
 //!     three_bits: u8,
 //!
 //!     /// Whole `u8`
-//!     #[wire(bytes = 1)]
+//!     // #[wire(bytes = 1)] Attribute not required as it is the default width of a `u8`
 //!     one_byte: u8,
 //!
 //!     /// Whole `u16`
-//!     #[wire(bytes = 2)]
+//!     // #[wire(bytes = 2)] Attribute not required as it is the default width of a `u16`
 //!     one_word: u16,
 //! }
 //! ```
@@ -90,8 +94,8 @@
 //!
 //! <section class="warning">
 //!
-//! It is the end user's responsibility to ensure the generic is of the correct size compared to
-//! the `#[wire(bits|bytes)]` attribute on the containing struct.
+//! It is the end user's responsibility to ensure the generic is of the correct size compared to the
+//! `#[wire(bits|bytes)]` attribute on the containing struct.
 //!
 //! </section>
 //!